[Layout - tips and techniques] Update description
Adding a new page description to tips and techniques so that it isnt a copy of the overview description:

Layout creates a context for communication on a page. Using the right composition helps guide a user through content in meaningful, useful and beautiful ways. The tips and techniques found here can help you organize and manipulate content for a successful user experience.